Refaktoroi ErityisenKoulutustehtävänJakso omaan tiedostoon
