+++ src/main/scala/fi/oph/koski/schema/Lukio.scala
@@ -59,20 +59,6 @@ case class LukionOpiskeluoikeudenLisätiedot(
   sisäoppilaitosmainenMajoitus: Option[List[Aikajakso]] = None
 ) extends OpiskeluoikeudenLisätiedot with ErityisenKoulutustehtävänJaksollinen
 
-@Description("Opiskelija opiskelee erityisen koulutustehtävän mukaisesti (ib, musiikki, urheilu, kielet, luonnontieteet, jne.). Kentän puuttuminen tai null-arvo tulkitaan siten, ettei opiskelija opiskele erityisen koulutustehtävän mukaisesti")
-case class ErityisenKoulutustehtävänJakso(
-  alku: LocalDate,
-  loppu: Option[LocalDate],
-  @Description("Erityinen koulutustehtävä. Koodisto")
-  @KoodistoUri("erityinenkoulutustehtava")
-  @OksaUri("tmpOKSAID181", "erityinen koulutustehtävä")
-  tehtävä: Koodistokoodiviite
-) extends Jakso
-
-trait ErityisenKoulutustehtävänJaksollinen {
-  val erityisenKoulutustehtävänJaksot: Option[List[ErityisenKoulutustehtävänJakso]]
-}
-
 trait LukionPäätasonSuoritus extends KoskeenTallennettavaPäätasonSuoritus with Toimipisteellinen with Suorituskielellinen
 
 @Description("Lukion oppimäärän suoritustiedot")
